sump-pump-serial/requirements.txt
pyserial>=3.0.0
redis>=4.0.0